[bc][log] Rename package log2 -> log
