+++ log/_examples/uselib/main.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 
 	// you can use package with _ prefix, but it is ignored by go build?
-	"github.com/dyweb/gommon/log2/_examples/uselib/foo"
-	"github.com/dyweb/gommon/log2/_examples/uselib/service"
-	"github.com/dyweb/gommon/log2/_examples/uselib/storage"
+	"github.com/dyweb/gommon/log/_examples/uselib/foo"
+	"github.com/dyweb/gommon/log/_examples/uselib/service"
+	"github.com/dyweb/gommon/log/_examples/uselib/storage"
 
-	_ "github.com/dyweb/gommon/log2/_examples/uselib/storage/mem"
+	_ "github.com/dyweb/gommon/log/_examples/uselib/storage/mem"
 )
 
 // TODO: make it a full example, like user auth with multiple mock backends
