+++ log/identity_test.go
@@ -1,4 +1,4 @@
-package log2
+package log
 
 import (
 	"fmt"
@@ -35,7 +35,7 @@ func TestNewPackageLogger(t *testing.T) {
 	assert.Equal(PackageLogger, id.Type)
 	assert.Equal("pkg", id.Type.String())
 	assert.Equal("init", id.Function)
-	assert.Equal("/home/at15/workspace/src/github.com/dyweb/gommon/log2/identity_test.go:10",
+	assert.Equal("/home/at15/workspace/src/github.com/dyweb/gommon/log/identity_test.go:10",
 		fmt.Sprintf("%s:%d", id.File, id.Line))
 }
 
@@ -54,7 +54,7 @@ func TestNewStructLogger(t *testing.T) {
 	assert.Equal("struct", id.Type.String())
 	assert.Equal("Foo", id.Struct)
 	assert.Equal(MagicStructLoggerFunctionName, id.Function)
-	assert.Equal("/home/at15/workspace/src/github.com/dyweb/gommon/log2/identity_test.go:22",
+	assert.Equal("/home/at15/workspace/src/github.com/dyweb/gommon/log/identity_test.go:22",
 		fmt.Sprintf("%s:%d", id.File, id.Line))
 }
 
@@ -68,6 +68,6 @@ func TestNewMethodLogger(t *testing.T) {
 	assert.Equal("method", id.Type.String())
 	assert.Equal("Foo", id.Struct)
 	assert.Equal("method", id.Function)
-	assert.Equal("/home/at15/workspace/src/github.com/dyweb/gommon/log2/identity_test.go:26",
+	assert.Equal("/home/at15/workspace/src/github.com/dyweb/gommon/log/identity_test.go:26",
 		fmt.Sprintf("%s:%d", id.File, id.Line))
 }
