+++ log/level.go
@@ -1,4 +1,4 @@
-package log2
+package log
 
 // Level is log level
 type Level uint8
