+++ log/level_test.go
@@ -1,4 +1,4 @@
-package log2
+package log
 
 import (
 	"fmt"
