+++ log/logger_test.go
@@ -1,1 +1,1 @@
-package log2
+package log
